refactor: ajuste para separação do main, do teste de contas
